requirements.txt
absl-py==2.1.0
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
comm==0.2.2
conda==22.11.1
conda-content-trust==0.1.3
conda-package-handling==1.9.0
contourpy==1.2.1
cryptography==38.0.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
eval_type_backport==0.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
fonttools==4.53.0
fqdn==1.5.1
fsspec==2024.6.0
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
imageio==2.37.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab-language-pack-zh-CN==4.2.post1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
MedPy==0.5.2
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
numpy==1.26.4
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pip==22.3.1
platformdirs==4.2.2
pluggy==1.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat==0.6.4
pycparser==2.21
pydantic==2.10.0
pydantic_core==2.27.0
Pygments==2.18.0
pyOpenSSL==22.0.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==65.5.0
SimpleITK==2.4.0
simsimd==6.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
stringzilla==3.10.10
supervisor==4.2.5
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.1.10
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.0
torch==2.1.2+cu118
torchvision==0.16.2+cu118
tornado==6.4.1
tqdm==4.64.1
traitlets==5.14.3
triton==2.1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==1.26.13
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.37.1
widgetsnbextension==4.0.11
